feat(walletconnect): code organization improvements
