+++ packages/app-dashboard/.env.example
@@ -3,6 +3,8 @@ VITE_VINCENT_DATIL_CONTRACT=
 VITE_STYTCH_PUBLIC_TOKEN=
 VITE_STYTCH_PROJECT_ID=
 
+VITE_WALLETCONNECT_PROJECT_ID=
+
 VITE_LIT_PAYER_SECRET_KEY=
 
 VITE_JWT_EXPIRATION_MINUTES=2160
