+++ packages/app-dashboard/public/.well-known/walletconnect.txt
@@ -1,1 +1,1 @@
-6ed5dae4-f9f6-41ce-8b05-95b6a17153b7=346002a05e7ce27f63d001b15bebb4fd4f9e49dcca55a6fa325f9c6f644c427d
+6ed5dae4-f9f6-41ce-8b05-95b6a17153b7=1905484c3d0b4e46e3f4b06d981aa7ebce8af29d2a38c78f0866c53fadc80c0a
