+++ packages/app-dashboard/src/config/env.ts
@@ -25,6 +25,7 @@ export const env = createEnv({
     VITE_VINCENT_DATIL_CONTRACT: z.string(),
     VITE_STYTCH_PUBLIC_TOKEN: z.string(),
     VITE_STYTCH_PROJECT_ID: z.string(),
+    VITE_WALLETCONNECT_PROJECT_ID: z.string(),
     VITE_LIT_PAYER_SECRET_KEY: z.string(),
     VITE_JWT_EXPIRATION_MINUTES: z.coerce.number(),
     VITE_GAS_BUFFER_DIVISOR: z.coerce.number(),
